Removed racey threading
As far as I can tell, the go-git library is not thread-safe and likes to write to objects while iterating through trees / commits. Because of this, we cannot easily thread the generation of html for each commit. This change reverts to a state where the threading steps around commit touching.