Finally removing recent commit
While finding the most recent commit for each file allowed us to not rewrite each one, it also requires us to run a diff on each file which is currently our most expensive operation. Since we are already doing this operation for each commit when we generate the log page's stats we will likely revisit this.